requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
dimod>=0.12.0
dwave-neal>=0.5.0
dwave-system>=1.20.0
scikit-learn>=1.3.0
openpyxl>=3.1.0

